Refactor campaign history display to separate component
