+++ components/Ads/CampaignBuilder.js
@@ -1,5 +1,5 @@
 import styles from './builder.module.css';
-import { Text, Box, Select, FormLabel, Container, FormControl, Textarea, Input, Button, TableContainer, Thead, Td, Tr, Th, Tbody, Table, Accordion, AccordionItem, AccordionButton, AccordionPanel } from "@chakra-ui/react";
+import { Text, Box, Select, FormLabel, FormControl, Textarea, Input, Button} from "@chakra-ui/react";
 import { useState, useRef } from 'react'
 const CURRENT_ADVERTISER = 'test';
